move opting functions to utils
