+++ src/smsapp.js
@@ -70,7 +70,7 @@ go.app = function() {
 
     // OPTOUT STATES
         self.states.add('state_opt_out_enter', function(name) {
-            return go.utils_project
+            return go.utils
                 .opt_out(self.im, self.contact)
                 .then(function() {
                     return self.states.create('state_opt_out');
@@ -87,7 +87,7 @@ go.app = function() {
 
     // OPTIN STATES
         self.states.add('state_opt_in_enter', function(name) {
-            return go.utils_project
+            return go.utils
                 .opt_in(self.im, self.contact)
                 .then(function() {
                     return self.states.create('state_opt_in');
