+++ src/utils_project.js
@@ -43,102 +43,6 @@ go.utils_project = {
     },
 
 
-// SUBSCRIPTION HELPERS
-
-    subscription_unsubscribe_all: function(contact, im) {
-        var params = {
-            'details__addresses__msisdn': contact.msisdn
-        };
-        return go.utils
-        .service_api_call("identities", "get", params, null, 'subscription/', im)
-        .then(function(json_result) {
-            // make all subscriptions inactive
-            var subscriptions = json_result.data;
-            var clean = true;  // clean tracks if api call is unnecessary
-            var patch_calls = [];
-            for (i=0; i<subscriptions.length; i++) {
-                if (subscriptions[i].active === true) {
-                    var updated_subscription = subscriptions[i];
-                    var endpoint = 'subscription/' + updated_subscription.id + '/';
-                    updated_subscription.active = false;
-                    // store the patch calls to be made
-                    patch_calls.push(function() {
-                        return go.utils.service_api_call("identities", "patch",
-                            {}, updated_subscription, endpoint, im);
-                    });
-                    clean = false;
-                }
-            }
-            if (!clean) {
-                return Q
-                .all(patch_calls.map(Q.try))
-                .then(function(results) {
-                    var unsubscribe_successes = 0;
-                    var unsubscribe_failures = 0;
-                    for (var index in results) {
-                        (results[index].code >= 200 && results[index].code < 300)
-                            ? unsubscribe_successes += 1
-                            : unsubscribe_failures += 1;
-                    }
-
-                    if (unsubscribe_successes > 0 && unsubscribe_failures > 0) {
-                        return Q.all([
-                            im.metrics.fire.inc(["total", "subscription_unsubscribe_success", "last"].join('.'), {amount: unsubscribe_successes}),
-                            im.metrics.fire.sum(["total", "subscription_unsubscribe_success", "sum"].join('.'), unsubscribe_successes),
-                            im.metrics.fire.inc(["total", "subscription_unsubscribe_fail", "last"].join('.'), {amount: unsubscribe_failures}),
-                            im.metrics.fire.sum(["total", "subscription_unsubscribe_fail", "sum"].join('.'), unsubscribe_failures)
-                        ]);
-                    } else if (unsubscribe_successes > 0) {
-                        return Q.all([
-                            im.metrics.fire.inc(["total", "subscription_unsubscribe_success", "last"].join('.'), {amount: unsubscribe_successes}),
-                            im.metrics.fire.sum(["total", "subscription_unsubscribe_success", "sum"].join('.'), unsubscribe_successes)
-                        ]);
-                    } else if (unsubscribe_failures > 0) {
-                        return Q.all([
-                            im.metrics.fire.inc(["total", "subscription_unsubscribe_fail", "last"].join('.'), {amount: unsubscribe_failures}),
-                            im.metrics.fire.sum(["total", "subscription_unsubscribe_fail", "sum"].join('.'), unsubscribe_failures)
-                        ]);
-                    } else {
-                        return Q();
-                    }
-                });
-            } else {
-                return Q();
-            }
-        });
-    },
-
-
-// OPTOUT & OPTIN HELPERS
-
-    opt_out: function(im, contact) {
-        contact.extra.optout_last_attempt = go.utils.get_today(im.config)
-            .format('YYYY-MM-DD hh:mm:ss.SSS');
-
-        return Q.all([
-            im.contacts.save(contact),
-            go.utils_project.subscription_unsubscribe_all(contact, im),
-            im.api_request('optout.optout', {
-                address_type: "msisdn",
-                address_value: contact.msisdn,
-                message_id: im.msg.message_id
-            })
-        ]);
-    },
-
-    opt_in: function(im, contact) {
-        contact.extra.optin_last_attempt = go.utils.get_today(im.config)
-            .format('YYYY-MM-DD hh:mm:ss.SSS');
-        return Q.all([
-            im.contacts.save(contact),
-            im.api_request('optout.cancel_optout', {
-                address_type: "msisdn",
-                address_value: contact.msisdn
-            }),
-        ]);
-    },
-
-
 // IDENTITY HELPERS
 
     find_healthworker_with_personnel_code: function(im, personnel_code) {
